requirements.txt
alembic==1.13.2
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
certifi==2024.7.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
psycopg==3.2.1
python-telegram-bot==21.4
sniffio==1.3.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
